% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all_plots.R
\name{all_plots}
\alias{all_plots}
\title{Generates all plots.}
\usage{
all_plots(biallelic = NULL, triallelic = NULL, genVars = NULL,
  treeInfo = NULL, config = NULL)
}
\arguments{
\item{biallelic}{A list called 'biallelic' created from the bugwas function}

\item{triallelic}{A list called 'triallelic' created from the bugwas function}

\item{genVars}{A list called 'genVars' created from the bugwas function}

\item{treeInfo}{A list called 'treeInfo' created from the bugwas function}

\item{config}{A list called 'config' created from the bugwas function}
}
\description{
This function generates all the plots
}
\examples{
all_plots()
}
\keyword{plot}

