% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianWaldTestPCsBarplot.R
\name{bayesianWaldTestPCsBarplot}
\alias{bayesianWaldTestPCsBarplot}
\title{Barplot of Bayesian Wald Test on principal components}
\usage{
bayesianWaldTestPCsBarplot(config, biallelic, treeInfo, colourPalette = NULL,
  p.genomewidepc = NULL)
}
\arguments{
\item{config}{A list called 'config' created from the lin_loc function. It is a required input.}

\item{biallelic}{A list called 'biallelic' created from the lin_loc function. It is a required input.}

\item{treeInfo}{A list called 'treeInfo' created from the lin_loc function. It is a required input.}

\item{colourPalette}{A vector of colours colour the significant principal components identified by the Bayesian Wald test (see testGenomeWidePCs). If this is NULL then colours are chosen from a default colour palette.}

\item{p.genomewidepc}{A matrix of the significant principal component and their correlation with lineages. This is Bayesian Wald test results produced by the function testGenomeWidePCs. If this is NULL then testGenomeWidePCs is called to generate the required test results.}
}
\description{
This function generates the barplot of Bayesian Wald Test on principal components.
}
\examples{
data <- lin_loc(gen = gen, pheno = pheno, phylo = phylo, 
 prefix = prefix, gem.path = gem.path)
testGenomeWidePCs(config = data$config, biallelic = data$biallelic)
}
\keyword{Barplot}
\keyword{Bayesian-Wald-test}
\keyword{PCA}

