% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataFromGemma.R
\name{getDataFromGemma}
\alias{getDataFromGemma}
\title{This function retrieves the binary data from GEMMA input files.}
\usage{
getDataFromGemma(prefix = NULL, gemmaGenFile = NULL, gemmaSnpFile = NULL,
  id = NULL)
}
\arguments{
\item{prefix}{Prefix of the output files.}

\item{gemmaGenFile}{The path to the GEMMA gen input file. It is a required input.}

\item{gemmaSnpFile}{The path to the GEMMA snp input file. It is a required input.}

\item{id}{A vector of the sample ids.}
}
\value{
A list of information on the biallelic snps
}
\description{
This function performs the genome-wide principal components.
}
\examples{
pheno <- read.table(file = pheno, header=T, as.is = T, sep="\\t")
data <- getDataFromGemma(gemmaGenFile = gemmaGenFile, gemmaSnpFile = gemmaSnpFile, 
							id = pheno.df$id, prefix = prefix)
}
\keyword{GEMMA}
\keyword{SNPs}

