% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSNPDataFromGEMMA.R
\name{getSNPDataFromGEMMA}
\alias{getSNPDataFromGEMMA}
\title{This function retrieves the binary data from GEMMA input files and phenotype files.}
\usage{
getSNPDataFromGEMMA(gemmaGenFile = NULL, gemmaSnpFile = NULL,
  pheno = NULL, prefix = NULL)
}
\arguments{
\item{gemmaGenFile}{The path to the GEMMA gen input file. It is a required input.}

\item{gemmaSnpFile}{The path to the GEMMA snp input file. It is a required input.}

\item{pheno}{The path of the phenotype file with contains sample ids in a column and phenotypes in another. The columns containing the ids and phenotypes have the repective headings id and phenotype.}

\item{prefix}{Prefix of the output files.}
}
\value{
A list of information on the biallelic snps and phenotypes.
}
\description{
This function generates the Manhattan plot(s) for a SNP GWAS.
}
\examples{
data <- getSNPDataFromGEMMA(gemmaGenFile = gemmaGenFile, gemmaSnpFile = gemmaSnpFile, 
							pheno = pheno, prefix = prefix)
}
\keyword{Data}
\keyword{GEMMA}

