% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logregVsLMM.R
\name{logregVsLMM}
\alias{logregVsLMM}
\title{Plot of the P-values of logistic regression versus those of LMM}
\usage{
logregVsLMM(config, biallelic, triallelic, colourPalette = NULL)
}
\arguments{
\item{config}{A list called 'config' created from the bugwas function. It is a required input.}

\item{biallelic}{A list called 'biallelic' created from the bugwas function. It is a required input.}

\item{triallelic}{A list called 'triallelic' created from the bugwas function. It is a required input.}

\item{colourPalette}{A vector of colours colour the significant principal components identified by the Bayesian Wald test (see testGenomeWidePCs). If this is NULL then colours are chosen from a default colour palette.}
}
\description{
This function generates the plot of the P-values of logistic regression versus those of SNP GWAS.
}
\examples{
logregVsLMM(biallelic = biallelic, triallelic = triallelic, config = config)
}
\keyword{SNP}
\keyword{Scatter}
\keyword{plot}

