% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotIndividualBy2PCs.R
\name{plotIndividualBy2PCs}
\alias{plotIndividualBy2PCs}
\title{Plot of the sample on the first two principal components}
\usage{
plotIndividualBy2PCs(biallelic = NULL, config = NULL)
}
\arguments{
\item{biallelic}{A list called 'biallelic' created from the bugwas function. It is a required input.}

\item{config}{A list called 'config' created from the bugwas function. It is a required input.}
}
\description{
This function generates a plot of the sample on the first two principal components.
}
\examples{
plotIndividualBy2PCs(biallelic = biallelic, config = config)
}
\keyword{Reduced}
\keyword{plot}
\keyword{space}

