% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpManhattanPlot.R
\name{snpManhattanPlot}
\alias{snpManhattanPlot}
\title{Generates Manhattan plots for a SNP GWAS}
\usage{
snpManhattanPlot(biallelic, triallelic, config, colourPalette = NULL)
}
\arguments{
\item{biallelic}{A list called 'biallelic' created from the bugwas function. It is a required input.}

\item{triallelic}{A list called 'triallelic' created from the bugwas function. It is a required input.}

\item{config}{A list called 'config' created from the bugwas function. It is a required input.}

\item{colourPalette}{A vector of colours colour the significant principal components identified by the Bayesian Wald test (see testGenomeWidePCs). If this is NULL then colours are chosen from a default colour palette.}
}
\description{
This function generates the Manhattan plot(s) for a SNP GWAS.
}
\examples{
snpManhattanPlot(biallelic = biallelic, triallelic = triallelic, config = config)
}
\keyword{Manhattan}
\keyword{SNP}
\keyword{plot}

