% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testGenomeWidePCs.R
\name{testGenomeWidePCs}
\alias{testGenomeWidePCs}
\title{Testing the genome-wide principal components}
\usage{
testGenomeWidePCs(config = NULL, biallelic = NULL)
}
\arguments{
\item{config}{A list called 'config' created from the bugwas function}

\item{biallelic}{A list called 'biallelic' created from the bugwas function}
}
\description{
This function performs the genome-wide principal components.
}
\examples{
testGenomeWidePCs(config, biallelic)
}
\keyword{Bayesian}
\keyword{Wald}
\keyword{components}
\keyword{principal}
\keyword{test,}

