% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testGenomeWidePCsAll.R
\name{testGenomeWidePCsAll}
\alias{testGenomeWidePCsAll}
\title{Testing the genome-wide principal components for all PCs.}
\usage{
testGenomeWidePCsAll(config = NULL, biallelic = NULL)
}
\arguments{
\item{config}{A list called 'config' created from the bugwas function}

\item{biallelic}{A list called 'biallelic' created from the bugwas function}
}
\value{
The p-value of the Bayesian Wald test for the genome-wide effect of principal components.
}
\description{
This function identifies the non-genome-wide principal components.
}
\examples{
data <- lin_loc(gen = gen, pheno = pheno, phylo = phylo,
 prefix = prefix, gem.path = gem.path)
testGenomeWidePCsAll(config = data$config, biallelic = data$biallelic)
}
\keyword{Bayesian-Wald-test}
\keyword{PCA}

