% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trueAndPredPhenoOnTreePlot.R
\name{trueAndPredPhenoOnTreePlot}
\alias{trueAndPredPhenoOnTreePlot}
\title{Generates all plots.}
\usage{
trueAndPredPhenoOnTreePlot(config, biallelic, treeInfo, p.genomewidepc = NULL,
  colourPalette = NULL)
}
\arguments{
\item{config}{A list called 'config' created from the lin_loc function. It is a required input.}

\item{biallelic}{A list called 'biallelic' created from the lin_loc function. It is a required input.}

\item{treeInfo}{A list called 'treeInfo' created from the lin_loc function. It is a required input.}

\item{p.genomewidepc}{A matrix of the significant principal component and their correlation with lineages. This is Bayesian Wald test results produced by the function testGenomeWidePCs. If this is NULL then testGenomeWidePCs is called to generate the required test results.}

\item{colourPalette}{A vector of colours colour the significant principal components identified by the Bayesian Wald test (see testGenomeWidePCs). If this is NULL then colours are chosen from a default colour palette.}
}
\description{
This function that plots the true and predicted phenotypes on a tree.
}
\examples{
data <- lin_loc(gen = gen, pheno = pheno, phylo = phylo, 
 prefix = prefix, gem.path = gem.path)
trueAndPredPhenoOnTreePlot(biallelic = data$biallelic, treeInfo = data$treeInfo,
 config = data$config)
}
\keyword{Phenotype}
\keyword{Phylogram}

