% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcLoadingPlot.R
\name{pcLoadingPlot}
\alias{pcLoadingPlot}
\title{Plot PC loadings of all SNPs.}
\usage{
pcLoadingPlot(config, biallelic)
}
\arguments{
\item{config}{A list called 'config' created from the lin_loc function. It is a required input.}

\item{biallelic}{A list called 'biallelic' created from the lin_loc function. It is a required input.}
}
\description{
This function plots the loadings of all SNPs for each significant principle component identified by the Bayesian Wald test.
}
\examples{
data <- lin_loc(gen = gen, pheno = pheno, phylo = phylo, 
 prefix = prefix, gem.path = gem.path)
pcLoadingPlot(config = data$config, biallelic = data$biallelic)
}
\keyword{Scatter-plot}

