% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndividualBy2PCs.R
\name{plotIndividualBy2PCs}
\alias{plotIndividualBy2PCs}
\title{Plot of the sample on the first two principal components}
\usage{
plotIndividualBy2PCs(biallelic = NULL, config = NULL)
}
\arguments{
\item{biallelic}{A list called 'biallelic' created from the lin_loc function. It is a required input.}

\item{config}{A list called 'config' created from the lin_loc function. It is a required input.}
}
\description{
This function generates a plot of the sample on the first two principal components.
}
\examples{
data <- lin_loc(gen = gen, pheno = pheno, phylo = phylo, 
 prefix = prefix, gem.path = gem.path)
plotIndividualBy2PCs(biallelic = data$biallelic, config = data$config)
}
\keyword{Reduced-space-plot}

