% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpPCManhattanPlot.R
\name{snpPCManhattanPlot}
\alias{snpPCManhattanPlot}
\title{Plot the a Manhattan plot organised by the signficance of PCs.}
\usage{
snpPCManhattanPlot(config, biallelic, triallelic, colourPalette = NULL)
}
\arguments{
\item{config}{A list called 'config' created from the lin_loc function. It is a required input.}

\item{biallelic}{A list called 'biallelic' created from the lin_loc function. It is a required input.}

\item{triallelic}{A list called 'triallelic' created from the lin_loc function. It is a required input.}

\item{colourPalette}{A vector of colours colour the significant principal components identified by the Bayesian Wald test (see testGenomeWidePCs). If this is NULL then colours are chosen from a default colour palette.}

\item{p.genomewidepc}{A matrix of the significant principal component and their correlation with lineages. This is Bayesian Wald test results produced by the function testGenomeWidePCs. If this is NULL then testGenomeWidePCs is called to generate the required test results.}
}
\description{
This function generates a Manhattan plot organised by the signficance of the principal components.
}
\examples{
data <- lin_loc(gen = gen, pheno = pheno, phylo = phylo, 
 prefix = prefix, gem.path = gem.path)
snpPCManhattanPlot(biallelic = data$biallelic,  triallelic = data$triallelic,
 config = data$config)
}
\keyword{Manhattan-plot}
\keyword{PCA}
\keyword{SNP}

